Dev update 6
- Testing robot actions:
	- New pre-built actions using the new CombinedAction mechanism.
- Working on vision system:
	- Fixed issue #8: ratio filtering now works within expected parameters
	- Vision runners now use the new dynamic vision system
	- Removed old vision system
	- Combined Hsv and Rgb filters into ColorFilter
- Robot Factory:
	- Added RobotState implementation for sbc
	- Added methods to check existence of axes and buttons
- BandwidthTracker:
	- Fixed bandwidth checker to return a double value
- Flashboard:
	- Log window now works without fxml and can show several logs.
	- Error dialogs are now shown when needed.
	- Integrated the new vision system
	- Changed the conversion path of an image received from the robot
	- Automatic loading of default vision filters added

+++ flashlibj/src/edu/flash3388/flashlib/communications/BandwidthTracker.java
@@ -86,10 +86,10 @@ public class BandwidthTracker implements CommInterface{
 		commInterface.update(millis);
 	}
 	
-	public long getBandwidthUsage(){
+	public double getBandwidthUsage(){
 		if(infoStart < 0)
 			return 0;
-		return (long) ((double)(bytes * 0.000008) / ((double)(FlashUtil.millis() - infoStart) * 0.001));
+		return ((double)(bytes * 0.000008) / ((double)(FlashUtil.millis() - infoStart) * 0.001));
 	}
 	public void reset(){
 		infoStart = -1;
